Remove EntityId class. Move its methods to Entity.
