+++ server/src/main/java/org/spine3/server/entity/Entity.java
@@ -20,6 +20,10 @@
 
 package org.spine3.server.entity;
 
+import com.google.common.base.Function;
+import com.google.common.base.Joiner;
+import com.google.common.collect.ImmutableSet;
+import com.google.protobuf.Descriptors;
 import com.google.protobuf.Message;
 import com.google.protobuf.Timestamp;
 import org.spine3.server.reflect.Classes;
@@ -32,9 +36,9 @@ import java.util.Map;
 
 import static com.google.api.client.util.Throwables.propagate;
 import static com.google.common.base.Preconditions.checkNotNull;
+import static com.google.common.collect.Iterables.transform;
 import static com.google.common.collect.Maps.newHashMap;
 import static com.google.protobuf.util.TimeUtil.getCurrentTime;
-import static org.spine3.server.entity.EntityId.checkType;
 
 /**
  * A server-side object with an identity.
@@ -77,6 +81,13 @@ public abstract class Entity<I, S extends Message> {
      */
     private static final int STATE_CLASS_GENERIC_INDEX = 1;
 
+    private static final ImmutableSet<Class<?>> SUPPORTED_ID_TYPES = ImmutableSet.<Class<?>>builder()
+            .add(String.class)
+            .add(Long.class)
+            .add(Integer.class)
+            .add(Message.class)
+            .build();
+
     private final I id;
 
     private S state;
@@ -95,7 +106,7 @@ public abstract class Entity<I, S extends Message> {
     public Entity(I id) {
         // We make the constructor public in the abstract class to avoid having protected constructors in derived
         // classes. We require that entity constructors be public as they are called by repositories.
-        checkType(id);
+        checkIdType(id);
         this.id = id;
     }
 
@@ -238,6 +249,62 @@ public abstract class Entity<I, S extends Message> {
     public static <I> Class<I> getIdClass(Class<? extends Entity<I, ?>> entityClass) {
         final Class<I> idClass = Classes.getGenericParameterType(entityClass, ID_CLASS_GENERIC_INDEX);
         return idClass;
+    }
+
+    /**
+     * Ensures that the type of the {@code entityId} is supported.
+     *
+     * @param entityId the ID of the entity to check
+     * @throws IllegalArgumentException if the ID is not of one of the supported types
+     */
+    public static <I> void checkIdType(I entityId) {
+        final Class<?> idClass = entityId.getClass();
+        if (SUPPORTED_ID_TYPES.contains(idClass)) {
+            return;
+        }
+        if (!Message.class.isAssignableFrom(idClass)){
+            throw unsupportedIdType(idClass);
+        }
+    }
+
+    /**
+     * Returns the short name of the ID type.
+     *
+     * @return
+     *  <ul>
+     *      <li>Short Protobuf type name if the value is {@link Message}.</li>
+     *      <li>Simple class name of the value, otherwise.</li>
+     *  </ul>
+     */
+    public String getShortIdTypeName() {
+        if (id instanceof Message) {
+            //noinspection TypeMayBeWeakened
+            final Message message = (Message) id;
+            final Descriptors.Descriptor descriptor = message.getDescriptorForType();
+            final String result = descriptor.getName();
+            return result;
+        } else {
+            final String result = id.getClass().getSimpleName();
+            return result;
+        }
+    }
+
+    private static IllegalArgumentException unsupportedIdType(Class<?> idClass) {
+        final String message = "Expected one of the following ID types: " + supportedTypesToString() +
+                "; found: " + idClass.getName();
+        throw new IllegalArgumentException(message);
+    }
+
+    private static String supportedTypesToString() {
+        final Iterable<String> classStrings = transform(SUPPORTED_ID_TYPES, new Function<Class<?>, String>() {
+            @Override
+            @SuppressWarnings("NullableProblems") // OK in this case
+            public String apply(Class<?> clazz) {
+                return clazz.getSimpleName();
+            }
+        });
+        final String result = Joiner.on(", ").join(classStrings);
+        return result;
     }
 
     /**
